third_party/blink/web_tests/external/wpt/html/semantics/document-metadata/interactions-of-styling-and-scripting/style-element-media-not-match-does-not-block-script.html

<!DOCTYPE html>
<title>Style element is not script-blocking when media doesn't match</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
const start = performance.now();
</script>
<style media="print">
@import url('stylesheet.py?delay=1');
</style>
<script>
// The loading of the imported sheet may start before the <style> element is set
// up due to preload scanner, so even when it's script-blocking, it may block
// script for a bit less than 1000ms. So we set an error margin to make the test
// more robust.
const epsilon = 50;
test(() => {
  const delay = performance.now() - start;
  assert_less_than(delay, 1000 - epsilon);
});
</script>
